f/18: testServerの起動をひとつにする。globalSetupとglobalTeardownを導入
